+++ src/Promotion/resources/manifest.php
@@ -4,5 +4,5 @@ declare(strict_types=1);
 
 return [
     'name' => 'Vanilo Promotion Module',
-    'version' => '4.2-dev'
+    'version' => '4.2.0'
 ];