+++ src/Shipment/Changelog.md
@@ -5,6 +5,7 @@
 ## Unreleased
 ##### 2024-XX-YY
 
+- Changed the minimal Address module requirement to v3.4.1
 - Changed `ShippingMethod::hasConfiguration()` to return false on empty arrays as well
 - Changed `Shipment::hasConfiguration()` to return false on empty arrays as well
 - Changed `Carrier::hasConfiguration()` to return false on empty arrays as well